added menu with react-router
